Remove functions from main.cpp and set them in main.h instead
Removing functions from main.cpp to make it more readable and easier to use and maintain
